prediction-service/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
Cartopy==0.22.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
cykhash==2.0.1
Cython==3.0.8
dadata==21.10.1
exceptiongroup==1.2.0
fastapi==0.108.0
fiona==1.9.5
fonttools==4.47.2
geopandas==0.14.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyogrio==0.7.2
pyparsing==3.1.1
pyproj==3.6.1
pyrobuf==0.9.3
pyrosm==0.6.2
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.14
pytz==2023.3.post1
requests==2.31.0
Rtree==1.1.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
pandas==2.1.4
scikit-learn==1.3.2
catboost==1.2.2
python-multipart==0.0.9
fastapi-cache2==0.2.1
redis==5.0.3




